move IsolatedLayerReader
