+++ hls4ml/converters/keras_v3/qkeras/__init__.py
@@ -1,1 +1,1 @@
-from . import activation, batchnorm, qdense, utils
+from . import activation, layer, utils
