+++ hls4ml/converters/keras_v3_to_hls.py
@@ -7,6 +7,7 @@ from typing import Any
 import numpy as np
 
 from hls4ml.model import ModelGraph
+from hls4ml.converters.utils import IsolatedLayerReader
 
 if typing.TYPE_CHECKING:
     import keras
@@ -238,15 +239,7 @@ class KerasV3HandlerDispatcher:
         config = layer.get_config()
         layer_dict = {'config': config, 'class_name': layer.__class__.__name__}
 
-        class IsolatedLayerReader:
-            def get_weights_data(self, layer_name, var_name):
-                assert layer_name == layer.name, f'Processing {layer.name}, but handler tried to read {layer_name}'
-                for w in layer.weights:
-                    if var_name in w.name:
-                        return np.array(w)
-                return None
-
-        reader = IsolatedLayerReader()
+        reader = IsolatedLayerReader(layer)
         input_shapes = [list(t.shape) for t in inp_tensors]
         input_names = [t.name for t in inp_tensors]
         output_names = [t.name for t in out_tensors]
